fix 2-d transform - inverse 2-d GPU still has a bug In the first implementation of the 2-d transform, I ran each row through all the transform steps, then each column. This version runs each row through 1 step, then each column through one step, then repeat on the upper left quadrant for further steps.

+++ data_io.cc
@@ -47,7 +47,7 @@ static bool readBinaryDataFile(const char *filename, float **data,
     return false;
   }
 
-  *data = new float[cells];
+  *data = new float[(int)cells];
   if (!*data) {
     fprintf(stderr, "Out of memory allocating %d x %d image\n",
             *width, *height);
@@ -55,7 +55,7 @@ static bool readBinaryDataFile(const char *filename, float **data,
     return false;
   }
 
-  fread(*data, 4, cells, inf);
+  fread(*data, 4, (int)cells, inf);
   fclose(inf);
 
   return true;
